requirements.txt
streamlit
graphviz
pillow